Add doc files and done changes according to review
