+++ content/docs/image/image-box-shadow.mdx
@@ -1,0 +1,33 @@
+---
+title: Image Box Shadow
+description: Brings box shadow on hover
+author: Sumin Gurung
+---
+
+<ComponentPreview name="image-image-box-shadow--docs" />
+
+# Installation
+
+1. Create a new file called `image-box-shadow.tsx` inside the `components/animata/image` directory.
+
+   macOS/Linux:
+
+   ```bash
+   mkdir -p components/animata/image && touch components/animata/image/image-box-shadow.tsx
+   ```
+
+   windows:
+
+   ```bash
+   mkdir components\animata\image && type null > components\animata\image\image-box-shadow.tsx
+   ```
+
+2. Open the newly created file and add the following code:
+
+```jsx file=<rootDir>/animata/image/image-box-shadow.tsx
+
+```
+
+# Credits
+
+Built by [Sumin Gurung](https://github.com/GrgSumin)
