+++ content/docs/image/zoom-image.mdx
@@ -1,0 +1,33 @@
+---
+title: Zoom Image
+description: Image will zoom in or out while hover
+author: Sumin Gurung
+---
+
+<ComponentPreview name="image-zoom-image--docs" />
+
+# Installation
+
+1. Create a new file called `zoom-image.tsx` inside the `components/animata/image` directory.
+
+   macOS/Linux:
+
+   ```bash
+   mkdir -p components/animata/image && touch components/animata/image/zoom-image.tsx
+   ```
+
+   windows:
+
+   ```bash
+   mkdir components\animata\image && type null > components\animata\image\zoom-image.tsx
+   ```
+
+2. Open the newly created file and add the following code:
+
+```jsx file=<rootDir>/animata/image/zoom-image.tsx
+
+```
+
+# Credits
+
+Built by [Sumin Gurung](https://github.com/GrgSumin)
